requirements.txt
fastapi==0.127.0
uvicorn==0.40.0
sqlalchemy==2.0.45
psycopg[binary]==3.3.2
python-dotenv==1.2.1
httpx==0.28.1
brotli-asgi==1.5.0
